Renaming namespaces.
