+++ src/Context.Tests/Context.Tests.csproj
@@ -1,8 +1,9 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.0</TargetFramework>
-
+    <AssemblyName>MongoDB.Extensions.Context.Tests</AssemblyName>
+    <RootNamespace>MongoDB.Extensions.Context.Tests</RootNamespace>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
@@ -22,10 +23,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Context\Context.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Update="Microsoft.SourceLink.GitHub" Version="1.0.0" />
   </ItemGroup>
 
 </Project>
