+++ src/Context/Internal/MongoCollectionBuilder.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using MongoDB.Bson.Serialization;
 using MongoDB.Driver;
 
-namespace MongoDB.Bootstrapper
+namespace MongoDB.Extensions.Context
 {
     internal class MongoCollectionBuilder<TDocument> : IMongoCollectionBuilder<TDocument>
     {
