+++ src/Context/Internal/MongoDatabaseBuilder.cs
@@ -6,7 +6,7 @@ using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Conventions;
 using MongoDB.Driver;
 
-namespace MongoDB.Bootstrapper
+namespace MongoDB.Extensions.Context
 {
     internal class MongoDatabaseBuilder : IMongoDatabaseBuilder
     {
